pms: Rename whole `conversations` subdirectory.
This completes a series replacing "conversation" as a jargon term
specific to PMs, in favor of "PM conversation".
